Seperated the one test file into two and removed the tests for Mocking.  Moved the helper classes into test_utils.py.
